requirements.txt
langchain
chromadb
pymupdf             # For PDF parsing
sentence-transformers
openai              # If using OpenAI for embeddings or LLM
streamlit           # Or fastapi, depending on your choice
python-dotenv
tqdm
langchain-community